Fixed a small issue when displaying SPARTAN frames. Now it will only display automatically if the frame is d3.
I am a little concerned with the problems of images that where not fully transferred and get displayed by half or so.  I will have to follow the program behavior to see what happens. Now, the program only display a frame if it is new to the database so, if SOI frames have complete header but no data, this may cause some issues.